requirements.txt
# Web Framework
Flask>=2.0.0
Flask-CORS>=4.0.0

# Machine Learning
scikit-learn>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
xgboost>=1.5.0
joblib>=1.1.0
openai>=1.0.0

# Data Processing
requests>=2.25.0
python-dotenv>=0.19.0
beautifulsoup4>=4.9.0

# Visualization
plotly>=5.0.0
matplotlib>=3.4.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.0.0

# Production
gunicorn>=20.1.0
python-json-logger>=2.0.0
sentry-sdk>=1.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Cache
redis>=6.0.0

# API
football-data-api==0.0.6